+++ packages/playground/src/components/filters/constants/conditions.ts
@@ -1,13 +1,13 @@
-import { FilterCondition } from './types'
+import type { ConditionOption } from '../types'
 
-const TYPE_CONDITIONS: FilterCondition[] = [
+export const BASIC_CONDITIONS: ConditionOption[] = [
   { label: 'is', value: 'is' },
   { label: 'is not', value: 'is_not' },
   { label: 'is empty', value: 'is_empty' },
   { label: 'is not empty', value: 'is_not_empty' }
 ]
 
-const DATE_CONDITIONS: FilterCondition[] = [
+export const RANGE_CONDITIONS: ConditionOption[] = [
   { label: 'is', value: 'is' },
   { label: 'is before', value: 'is_before' },
   { label: 'is after', value: 'is_after' },
@@ -15,5 +15,3 @@ const DATE_CONDITIONS: FilterCondition[] = [
   { label: 'is empty', value: 'is_empty' },
   { label: 'is not empty', value: 'is_not_empty' }
 ]
-
-export { TYPE_CONDITIONS, DATE_CONDITIONS }
