+++ packages/playground/src/types/repository.ts
@@ -1,0 +1,13 @@
+interface Repository {
+  id: string
+  name: string
+  description?: string
+  private: boolean
+  stars: number
+  forks: number
+  pulls: number
+  createdAt: string
+  timestamp: string
+}
+
+export type { Repository }
